requirements.txt
attrs==22.2.0
aws-cdk-lib==2.67.0
aws-cdk.asset-awscli-v1==2.2.102
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.82
black==23.1.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constructs==10.1.275
exceptiongroup==1.1.0
idna==3.4
importlib-resources==5.12.0
jsii==1.77.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
publication==0.0.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
